sys_core_fold: Improve case optimization
The optimization that avoids building a tuple in a case
expression would not work if any clause matched a tuple
as in the following example:

f(A, B) ->
    case {A,B} of
	{<<X>>,Y} ->
	    {X,Y}
    end.

The generated Core Erlang code would look like this (note
the tuples in the case expression and the pattern):

'f'/2 =
    fun (_cor1,_cor0) ->
	case {_cor1,_cor0} of
	  <{#{#<X>(8,1,'integer',['unsigned'|['big']])}#,Y}>
            when 'true' ->
	      {X,Y}
          .
          .
          .
	end

It is expected that the code should look like this (note
that tuples have been replaced with "values"):

'f'/2 =
    fun (_cor1,_cor0) ->
	%% Line 5
	case <_cor1,_cor0> of
	  <#{#<X>(8,1,'integer',['unsigned'|['big']])}#,Y> ->
	      {X,Y}
          .
          .
          .
	end

While at it, also fix bugs in the handling of pattern with
aliases. The bindings were produced in the wrong order (creating
'let's with referring to free variables), but in most cases
the incorrect bindings were discarded later without causing any
harm.