Let file:write_file/3 use writev
Previously, this function would turn any input into a single binary
before writing. This meant it could not take advantage of the `writev`
system call if it was given a list of binaries and told to write with
`raw` mode.

To see this, start an erlang shell on the parent commit, and also
start this dtrace script:

    https://github.com/evanmiller/tracewrite

like this:

    sudo dtrace -s tracewrite.d -p $(pgrep beam)

In the erlang shell, run the following:

    file:write_file("/tmp/tmp.txt", [<<97,98>>, <<98,97>>], [raw]).

In the dtrace output, you will see that the system call used is `write`.

Now repeat with this commit, and you will see that `writev` is used.