sys_core_fold: Run optimizations to a fixpoint
Run the optimizations until a fixpoint is reached, or until
the maximum iteration count is reached.

The hope is that in the future we can many small optimizations
instead of optimizations that try to do everything in one pass.

This change allows us to remove the ad-hoc calls to expr/2
to run more optimizations on a piece of code.