Revert "Check libsctp for sctp funcs in configure.in"
This reverts commit 52fae83743a6e84d719f4f82fe45e6e0efdbd6f0:

    Check libsctp for sctp funcs in configure.in

    Call AC_CHECK_LIB before calling AC_CHECK_FUNCS to check for functions in libsctp.
    Otherwise AC_CHECK_FUNCS will not link with libsctp to see if function exists.

Reverting to the old behaviour of not checking for libsctp before checking for
sctp functions.  Then it works on Linux by loading libsctp and looking up the
symbols i runtime, and it works on FreeBSD since there is no separate libsctp
to link against - the functions are part of the default system libraries.