dbg:fun2ms: allow empty list as head
Running 'dbg:fun2ms(fun([]) -> return_trace() end' resulted in an error
"dbg:fun2ms requires fun with single variable or list parameter"

But the empty list is actually a list and it is a valid value as a
match-spec head (matching on arity-0 functions).

Although its practical use is questionable this commit eliminates a
small limitation of ms_transform which is not present in the match-spec
grammar.