Code cleanup to remove confusing assertion
The code of add_ref/6 contained an assertion that somehow appeared
twice.  Also, a recent commit (2b2de61) suppressed an unmatched
return warning in a non-optimal way.  By changing the return type
of hipe_bifs:add_ref/2 to be 'ok' instead of [] the code can be
cleaned up and simplified.

While at it, fixed the wrong indentation of some nearby code.