Eliminate unmatched return dialyzer warnings
The code contained a lot of unmatched return warnings, mostly in function
calls that were executed only for their side-effects.  Also, there were
various places where possible errors were not checked and now they are.
This may cause crashes, both in the ftp testsuite (which I have not run)
and in situations were errors occur but are ignored.