Allow only X :: Y constraints after when in specs
The parser, most likely due to once upon a time supporting
is_subtype(_, _) constraints after 'when' in specs, allowed
'when's to be followed by all sorts of atom(...) constructs
and not only by type constraints of the X :: Y form.  Since
is_subtype/2 constraints are no longer allowed, the relevant
clause that parses them can now be taken out.

This change means that specs with a 'when' syntax which is
syntactically erroneous will now result in syntax errors,
instead of being passed to erl_lint or other parts of the
system and be found erroneous there.