Fix performance bug in erl_prim_loader:get_modules/{2,3}
The erl_prim_loader:get_modules/{2,3} functions were introduced in
cd283583f8. Unfortunately, while the functions worked correctly, there
was a bug in that many garbage maessages would be sent to the
erl_prim_loader process.  The number of extra messages depended on
both the length of the code path and of the number of modules that
were fetched.  The messages were ignored and ultimately discarded,
causing no harm except for a performance degradation and increase of
the heap size for the erl_prim_loader process.

The following functions were hit by the performance bug:

  code:atomic_load/1
  code:ensure_modules_loaded/1
  code:prepare_loading/1