Run dialyzer as part of the travis script
Build a dialyzer PLT and use it to analyze all OTP applications
that can currently be analyzed without warnings even when the
option -Wunmatched_returns is turned on. Note that the dialyzer
run does _not_ enable the option which allows for improper lists.

Applications to run dialyzer on are mentioned alphabetically.
As more applications are fixed to run cleanly even with unmatched
returns, they can be added to this list.  However, there will come
a point when the warning pass of Dialyzer will run out of memory
on Travis and the process will be killed.  This should be fixed in
dialyzer.