ssl: Correct terminate behaviour
When the terminate function is called explicitly, to make guarantees
that for instance the reuseaddr option works as expected, we must
make sure that the clean up code is not run again when gen_statem
calls terminate. This check was broken in the rewrite from gen_fsm to
gen_statem.

Caused PEM cache errors, that in some cases would
cause unexpected connection failures.