compile: Reduce memory consumption during compilation
The compiler would keep the data structures for two compiler
passes in memory. That could increase the maximum amount of
memory that the compiler uses, and could also have a negative
impact on performance (terms that would not be used again would be
copied by a garbage collection).

Here is an example that shows how the previous version of the code
could get captured:

a_compiler_pass(Mod, St) ->
    case Mod:module(St#compile.code, St#compile.options) of
      {ok,Code} ->
	{ok,St#compile{code=Code}};
    ...

The reference to the code from the previous pass will only be released
when St is updated. We can avoid the problem by passing the
current version of the code as a function argument:

a_compiler_pass(Mod, Code0, St) ->
    case Mod:module(Code0, St#compile.options) of
      {ok,Code} ->
	{ok,Code,St};
    ...

In practice, this change does not seem to significantly speed up
the compiler, but it does not do any harm either. It should help
dialyzer in situations when dialyzer compiles several large modules
at the same time.