add ErlNifEntry destructor
Adds dtor and dtor_data members to ErlNifEntry.  When the shared library containing the NIF module is unloaded from the system, a non-NULL dtor
will be invoked with dtor_data.  The purpose of this destructor is to free a heap-allocated ErlNifEntry and its associated arrays and strings.

The use case for this is NIF modules written in Rust or any other language that has difficulty forming static C strings and structures.  The
existing options in Rust are to use awkward and error prone constructs for a static solution as seen in
[erlang_nif-sys](https://github.com/goertzenator/erlang_nif-sys), or use heap-allocated data and accept that they will leak when unloaded as seen
in [rustler](https://github.com/hansihe/Rustler).

NIF modules written in C will always have a NULL destructor because it's ErlNifEntry is always static.