Emacs: Consider arity when jumping to definitions
Only the xref front-end introduced in Emacs 25 consider arity.  It is
not implemented for older emacsen.

Look for manual page files in lib/erlang/man in erlang-root-dir.  Also
do not give up in erlang-man-module when not finding the manual page
file.  Call manual-entry as a fallback.

Do not bother to populate menu-bar with man pages when menu-bar-mode
is nil.

Add erlang extensions also to dired-omit-extensions.

Remove some support for Emacs 18 and 19.