dialyzer: Fix check of parameterized opaque types
Two parameters used to be "compatible" if one is a specialization of
the other, in the way that the input type of t_limit() is a
specialization of the output type. This check is now relaxed: any() in
any of the two parameters is compatible with any type of the other
parameter.

The change is due to bugus warnings.

A clause for maps has been added.