Clean up handling of textual order
The to_encoding_order/1 function can be eliminated if we
incorporate its functionality into textual_order/2.

textual_order/2 has a workaround for TermList being longer
than OrderList. Remove the workaround, because the code
being generated would certainly be wrong (better let the compiler
crash and receive a bug report if it happens). The workaround
was not necessary to successfully compile the entire Erlang/OTP
and to run the asn1 test suite.