dialyzer: Sort graphs topologically
Although some variable names indicate that lists of SCCs and modules
are sorted, that was not always the case. The effect on the execution
time of sorting them doesn't seem to be significant.

dialyzer_coordinator:sccs_to_pids() should now always return an empty
second list (not yet started workers).

Since the condensation of graphs often needs a lot of heap memory, it
is run in a separate process.