[ct] Add Suite argument to hook callback functions
An extra argument, Suite, is added as the first argument to each of
the following hook callback functions:

  - pre_init_per_group
  - post_init_per_group
  - pre_end_per_group
  - post_end_per_group
  - pre_init_per_testcase
  - post_init_per_testcase
  - pre_end_per_testcase
  - post_end_per_testcase
  - on_tc_fail
  - on_tc_skip

For backwards compatibility, if the new function is not exported from
a hook callback module, common_test will fall back to the old
interface and call the function without the Suite argument.

The reason for adding the new argument is that if a test suite is
skipped by a 'skip_suites' statement in the test specification, then
there will be no call to pre/post_init_per_suite, and thus the hook
has no other way of knowing which Suite is skipped when it gets the
on_tc_skip callback. The other callbacks are updated for symmetry.