Fix observer application crash (#1296)
Fix observer application crash

When clicking an HTML-link to a port before the port tab has been opened for the first time, observer would crash since port info is not initiated. This is now corrected.

Also, when clicking on an HTML link to a port, and the port does not exist,
then pop up an info dialog saying "No such port". 

OTP-14151